- Запихнул новые модели - Запихнул новые анимации

+++ Assets/Scripts/SearchComponents/BaseTagMB.cs
@@ -6,7 +6,7 @@ public class BaseTagMB : MonoBehaviour
     public bool isFriendly; // to do ay del this after create baseSpawnSystems
 
     [Range(1, 10)]
-    public float TimeToSpawn = 5;
+    public int TimeToSpawn = 5;
 
     [Range(1, 4)]
     public int MonsterLevel;
